requirements.txt
streamlit
pytube 
openai-whisper 
google-generativeai 
python-dotenv